# lwip: socket: Add support for TCP_NODELAY.