This adds support to use the Nagle algorithm implemented already in LwIP to determine when a TCP should be sent.

I was confused as to why there was a constant ~200ms lag when sending small packets from TCP in Micropython. After too much searching, I arrived at the TCP_TMR_INTERVAL which sends packets at its usual timeout interval of 250ms. Changing this to 50ms makes a corresponding difference in the lag but still doesn't offer a solid solution on reducing the latency when small packets are desired.

I looked at adding the TCP_NODELAY socket option for LwIP, but that also didn't work until I found that it isn't used without calling `tcp_output_nagle`. And Micropython currently has a static algorithm to issue packet sends with the send buffer is below 25%. 

It seems to me we should be using the common Nagle algorithm to determine when best to send packets based on best performance / latency? I'm of course open to other ideas and happy to provide some before+after tests based on performance and/or latency if interested.

Maybe not said enough, thanks to all who have put together such an incredible project!