@greezybacon 2024-03-19
@dpgeorge sorry if my story-telling got the best of the situation.

250ms is too long. 50ms is also too long. If the controller provides a small response to a small message/request, it should be sent immediately- or at least be able to.