@dpgeorge 2024-03-19
> If the controller provides a small response to a small message/request, it should be sent immediately- or at least be able to.

Yes, I agree.  But it's hard to balance that requirement against good throughput when there is lots of data being written in small chunks.

Well, that's what this PR/discussion is about!