@greezybacon 2024-03-19
But that’s the reason for the no-delay “option”. It allows the application designer to make the decision on throughput / latency for the IP stack.