@greezybacon 2024-03-19
[Nagle’s algorithm](https://en.wikipedia.org/wiki/Nagle%27s_algorithm) is a well studied, general solution to the latency / throughput problem. It was published in RFC 896.

I don’t pretend to be an expert- just want to provide context:

When using it, it determines that packets should NOT be sent immediately if there is unacknowledged data previously sent. This allows the application to buffer data until either (a) the send buffer exceeds the max-segment-size (MSS) or (b) the receiver acknowledges all previously-sent packets.

This allows the sender to buffer writes using a dynamic threshold based on the latency of the receiver.

As previously written, Micropython would only send packets if there was <25% remaining in the send buffer. Using it, sending a small message of ~50 bytes would not trigger a send. So it would remained queued until the TCP interval timer expires.

Using Nagle’s algorithm, the first write, no matter how small, will generate a packet. Until that data is acknowledged, all further writes will be queued until one of the two criteria is met. Additionally, when the ACK is received from the receiver, it will trigger the queued data to be sent. So, if the MSS has not yet been reached by the application, then the maximum throughput of the application is still reached along with a reasonable latency given the delay distance of the receiver. If the MSS is reached by the app in the meantime, packets may be sent until the TCP window size remains unacknowledged.

In addition to this, an application designer may choose to always create packets for every write by setting the TCP_NODELAY socket option. But even without that option, Nagle’s algorithm will provide much more reasonable latency from Micropython.