@greezybacon 2024-03-30
@dpgeorge I wanted to provide a somewhat scientific approach to testing the throughput/latency problem. I constructed these two programs to setup a simple character-generator and echo-service to test the send throughput and response latency respectively. The character generator service sends 16-byte packets continuously to a connected client. The echo-service will read a 16-byte message and then respond with the received message.

Results on RP2040 with WIZnet W5500
-----

|build|character generator|echo service|
|------|------|-------|
| v1.23.0-preview.237.g7dff38fdc | 371 kiB/s | 21 msg/s  (~48ms response time)|
| This PR (without TCP_NODELAY) | 357 kiB/s | 240 msgs/s (~4.2ms response time) |
|------|------|-------|
| Diff | 96.2% (-3.77%) | 1143% |


So, to your point, it does impact the maximal throughput of an application that makes many small writes. However, that impact is rather small considering that for message-based applications it makes an 11 fold increase in the performance 
and response time.

------

Server:
```python
import socket
def network_service(handler, port=8888, backlog=2):
    host = socket.getaddrinfo('0.0.0.0', port)[0]
    s = socket.socket()
    s.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
    s.bind(host[-1])
    s.listen(backlog)
    #
    while True:
        try:
            s2, addr = s.accept()
            handler(s2)
        except KeyboardInterrupt:
            s.close()
            break

def chargen(s):
    data = b'a' * 16
    try:
        while True:
            for char in range(33, 65):
                s.write(data)
    except KeyboardInterrupt:
        raise
    finally:
        print("Client disconnected")
        s.close()

def echo(s):
    empty = b''
    try:
        while True:
            buf = s.read(16)
            if buf == empty:
                break
            s.write(buf)
    except KeyboardInterrupt:
        raise
    finally:
        print("Client disconnected")
        s.close()

import network
nic = network.WIZNET5K()
nic.active(True)
nic.ifconfig('dhcp')
print(f"Network up at {nic.ifconfig()[0]}")

network_service(chargen, port=8888)
```

And the client (running CPython)
```python
async def downloader(host, port, timeout=10):
    reader, writer = await asyncio.open_connection(host, port)
    #
    start = time.monotonic()
    end = start + timeout
    bytes = 0
    try:
        while time.monotonic() < end:
            bytes += len(await reader.read(512))
    finally:
        now = time.monotonic()
        avg = bytes / (now - start)
        print(f"Read {bytes >> 10}kiB bytes (avg {int(avg) >> 10}kiB/s) in {now - start:.1f}s")
        writer.close()

async def echo(host, port, timeout=10):
    reader, writer = await asyncio.open_connection(host, port)
    #
    start = time.monotonic()
    end = start + timeout
    msgs = 0
    try:
        while time.monotonic() < end:
            writer.write(b'a' * 16)
            await writer.drain()
            await reader.read(16)
            msgs += 1
    finally:
        now = time.monotonic()
        avg = msgs / (now - start)
        print(f"Read {msgs} msgs (avg {int(avg)}msgs/s) in {now - start:.1f}s")
        writer.close()
```

For the `chargen` service, I got these results on a RP2040 with WIZnet W5500:

This PR:
```
>>> asyncio.run(downloader(mphost, 8888))
Read 3229kiB bytes (avg 358kiB/s) in 10.0s
>>> asyncio.run(downloader(mphost, 8888))
Read 3238kiB bytes (avg 359kiB/s) in 10.0s
>>> asyncio.run(downloader(mphost, 8888))
Read 3206kiB bytes (avg 356kiB/s) in 10.0s
```

v1.23.0-preview.237.g7dff38fdc
```
>>> asyncio.run(downloader(mphost, 8888))
Read 3316kiB bytes (avg 368kiB/s) in 10.0s
>>> asyncio.run(downloader(mphost, 8888))
Read 3421kiB bytes (avg 380kiB/s) in 10.0s
>>> asyncio.run(downloader(mphost, 8888))
Read 3355kiB bytes (avg 372kiB/s) in 10.0s
```

For the `echo` service:

This PR:
```
>>> asyncio.run(echo(mphost 8888))
Read 1984 msgs (avg 220msgs/s) in 0.9s
>>> asyncio.run(echo(mphost, 8888))
Read 2358 msgs (avg 262msgs/s) in 0.9s
>>> asyncio.run(echo(mphost, 8888))
Read 2194 msgs (avg 243msgs/s) in 0.9s
>>> asyncio.run(echo(mphost, 8888))
Read 2140 msgs (avg 237msgs/s) in 0.9s
```

v1.23.0-preview.237.g7dff38fdc
```
>>> asyncio.run(echo(mphost, 8888))
Read 186 msgs (avg 20msgs/s) in 0.4s
>>> asyncio.run(echo(mphost, 8888))
Read 190 msgs (avg 21msgs/s) in 0.5s
>>> asyncio.run(echo(mphost, 8888))
Read 204 msgs (avg 22msgs/s) in 0.0s
```