@greezybacon 2024-05-23
@dpgeorge if it would help, I'd be happy to remove the TCP_NODELAY option.