@dpgeorge 2024-05-23
This is a very good improvement for short TCP messages, while still maintaining good link efficiency for lots of data.  And I think it's useful to have the TCP_NODELAY option so users have some control over how TCP output works, which can be important for IoT devices.

So, this is merged!