@vshymanskyy 2024-05-23
Now, it should be fairly easy to port it to ESP32.
Is it planned or a PR would be welcome?