@greezybacon 2024-05-23
@vshymanskyy I believe this code will work consistently for any NIC which is using lwIP stack.