@vshymanskyy 2024-05-23
Yeagh but ESP32 uses a different socket implementation:
https://github.com/micropython/micropython/blob/master/ports/esp32/modsocket.c

So I guess it needs to be adjusted separately